+++ tests/autosave_test.sh
@@ -1,9 +1,9 @@
 jq -n --argjson n 50 '[{"N": "a"}] + [range(1; $n + 1) | {N: .}]' > datas.json
 # cargo build --release
-./target/release/parabuild tests/example_run_time_consuming_project build/main --template-file src/main.cpp --data-file datas.json -j 2 -J 2 --output-file output.json > /dev/null 2>&1 &
+./target/release/parabuild tests/example_run_time_consuming_project build/main --template-file src/main.cpp --data-file datas.json -j 2 -J 2 --output-file output.json --autosave 1s > /dev/null 2>&1 &
 PID=$!
 sleep 10
-kill -SIGINT $PID
+kill -SIGKILL $PID
 wait $PID
 ./target/release/parabuild tests/example_run_time_consuming_project build/main --template-file src/main.cpp --data-file datas.json -j 2 -J 2 --output-file output.json --continue > /dev/null 2>&1
 run_len=$(jq length output.json)